refactor(#164): 리뷰 관련 컴포넌트 폴더 생성
